+++ examples/ftp/client/CMakeLists.txt
@@ -2,19 +2,14 @@ add_executable(w5x00_ftp_client
         w5x00_ftp_client.c
         )
 
-target_include_directories(w5x00_ftp_client PUBLIC
-        ${CMAKE_SOURCE_DIR}/libraries/ioLibrary_Driver/Ethernet
-        ${CMAKE_SOURCE_DIR}/libraries/ioLibrary_Driver/Internet/FTPClient
-		${CMAKE_CURRENT_SOURCE_DIR}
-        )
-
-target_link_libraries(w5x00_ftp_client PUBLIC
+target_link_libraries(w5x00_ftp_client PRIVATE
         pico_stdlib
         hardware_spi
         hardware_dma
         hardware_clocks
         ETHERNET_FILES
         FTPCLIENT_FILES
+        SPI_FILES
         )
 
 pico_enable_stdio_usb(w5x00_ftp_client 1)
