+++ examples/ftp/server/CMakeLists.txt
@@ -2,19 +2,14 @@ add_executable(w5x00_ftp_server
         w5x00_ftp_server.c
         )
 
-target_include_directories(w5x00_ftp_server PUBLIC
-        ${CMAKE_SOURCE_DIR}/libraries/ioLibrary_Driver/Ethernet
-        ${CMAKE_SOURCE_DIR}/libraries/ioLibrary_Driver/Internet/FTPServer
-		${CMAKE_CURRENT_SOURCE_DIR}
-        )
-
-target_link_libraries(w5x00_ftp_server PUBLIC
+target_link_libraries(w5x00_ftp_server PRIVATE
         pico_stdlib
         hardware_spi
         hardware_dma
         hardware_clocks
         ETHERNET_FILES
         FTPSERVER_FILES
+        SPI_FILES
         )
 
 pico_enable_stdio_usb(w5x00_ftp_server 1)
