+++ examples/loopback/README.md
@@ -27,7 +27,7 @@ If you are using W5100S-EVB-Pico, you can skip '1. Combine...'
 
 To test the Loopback example, minor settings shall be done in code.
 
-1. Setup SPI port and pin.
+1. Setup SPI port and pin in 'RP2040-HAT-C/port/ioLibrary_Driver/w5x00spi.h' directory.
 
 Setup the SPI interface you use.
 
@@ -49,7 +49,7 @@ If you want to test with the Loopback example using SPI DMA, uncomment USE_SPI_D
 //#define USE_SPI_DMA // if you want to use SPI DMA, uncomment.
 ```
 
-2. Setup network configuration such as IP.
+2. Setup network configuration such as IP in 'RP2040-HAT-C/examples/loopback/w5x00_loopback.c' directory.
 
 Setup IP and other network settings to suit your network environment.
 
@@ -66,7 +66,7 @@ static wiz_NetInfo g_net_info =
 };
 ```
 
-3. Setup loopback configuration.
+3. Setup loopback configuration in 'RP2040-HAT-C/examples/loopback/w5x00_loopback.c' directory.
 
 ```cpp
 /* Port */
@@ -120,8 +120,9 @@ Link
 [link-tera_term]: https://osdn.net/projects/ttssh2/releases/
 [link-hercules]: https://www.hw-group.com/software/hercules-setup-utility
 [link-raspberry_pi_pico_usb_mass_storage]: https://github.com/Wiznet/RP2040-HAT-C/blob/main/static/images/loopback/raspberry_pi_pico_usb_mass_storage.png
-[link-connect_to_serial_com_port]: https://github.com/Wiznet/RP2040-HAT-C/blob/main/static/images/loopback/connect_to_serial_com_port.png
+[link-connect_to_serial_com_port]: https://github.com/Wiznet/RP2040-HAT-C/main/static/images/loopback/connect_to_serial_com_port.png
 [link-see_network_information_of_raspberry_pi_pico_and_open_loopback_server]: https://github.com/Wiznet/RP2040-HAT-C/blob/main/static/images/loopback/see_network_information_of_raspberry_pi_pico_and_open_loopback_server.png
 [link-connect_to_loopback_server_using_hercules_tcp_client_1]: https://github.com/Wiznet/RP2040-HAT-C/blob/main/static/images/loopback/connect_to_loopback_server_using_hercules_tcp_client_1.png
 [link-connect_to_loopback_server_using_hercules_tcp_client_2]: https://github.com/Wiznet/RP2040-HAT-C/blob/main/static/images/loopback/connect_to_loopback_server_using_hercules_tcp_client_2.png
 [link-receive_back_sent_message]: https://github.com/Wiznet/RP2040-HAT-C/blob/main/static/images/loopback/receive_back_sent_message.png
+
