+++ examples/mqtt/CMakeLists.txt
@@ -1,2 +1,3 @@
 add_subdirectory(publish)
+add_subdirectory(publish_subscribe)
 add_subdirectory(subscribe)
