+++ examples/mqtt/publish/CMakeLists.txt
@@ -2,17 +2,13 @@ add_executable(w5x00_mqtt_publish
         w5x00_mqtt_publish.c
         )
 
-target_include_directories(w5x00_mqtt_publish PUBLIC
-        ${CMAKE_SOURCE_DIR}/libraries/ioLibrary_Driver/Ethernet
-        ${CMAKE_SOURCE_DIR}/libraries/ioLibrary_Driver/Internet/MQTT
-        )
-
-target_link_libraries(w5x00_mqtt_publish PUBLIC
+target_link_libraries(w5x00_mqtt_publish PRIVATE
         pico_stdlib
         hardware_spi
         hardware_dma
         ETHERNET_FILES
-        MQTT_FILES
+        MQTT_FILES        
+        SPI_FILES
         )
 
 pico_enable_stdio_usb(w5x00_mqtt_publish 1)
