+++ examples/sntp/CMakeLists.txt
@@ -2,17 +2,14 @@ add_executable(w5x00_sntp
         w5x00_sntp.c
         )
 
-target_include_directories(w5x00_sntp PUBLIC
-        ${CMAKE_SOURCE_DIR}/libraries/ioLibrary_Driver/Ethernet
-		${CMAKE_SOURCE_DIR}/libraries/ioLibrary_Driver/Internet/sntp
-        )
-
 target_link_libraries(w5x00_sntp PUBLIC
         pico_stdlib
         hardware_spi
         hardware_dma
         ETHERNET_FILES
-		SNTP_FILES
+        SNTP_FILES
+        SPI_FILES
+        TIMER_FILES
         )
 
 pico_enable_stdio_usb(w5x00_sntp 1)
