+++ examples/tcp_client_over_ssl/README.md
@@ -27,7 +27,7 @@ If you are using W5100S-EVB-Pico, you can skip '1. Combine...'
 
 To test the TCP Client over SSL example, minor settings shall be done in code.
 
-1. Setup SPI port and pin.
+1. Setup SPI port and pin in 'RP2040-HAT-C/port/ioLibrary_Driver/w5x00spi.h' directory.
 
 Setup the SPI interface you use.
 
@@ -49,7 +49,7 @@ If you want to test with the TCP Client over SSL example using SPI DMA, uncommen
 //#define USE_SPI_DMA // if you want to use SPI DMA, uncomment.
 ```
 
-2. Setup network configuration such as IP.
+2. Setup network configuration such as IP in 'RP2040-HAT-C/examples/tcp_client_over_ssl/w5x00_tcp_client_over_ssl.c' directory.
 
 Setup IP and other network settings to suit your network environment.
 
@@ -66,7 +66,7 @@ static wiz_NetInfo g_net_info =
 };
 ```
 
-3. Setup TCP Client over SSL configuration.
+3. Setup TCP Client over SSL configuration in 'RP2040-HAT-C/examples/tcp_client_over_ssl/w5x00_tcp_client_over_ssl.c' directory.
 
 In the TCP client over SSL configuration, the target IP is the IP of your desktop or laptop where SSL server will be created.
 
@@ -78,6 +78,8 @@ static uint8_t g_ssl_target_ip[4] = {192, 168, 11, 3};
 ```
 
 In order to change SSL settings, modify 'ssl_config.h' located in the same directory with TCP Client over SSL example.
+
+.
