--- patches/03_mbedtls_test_mode.patch
@@ -1,37 +1,0 @@
-diff --git a/CMakeLists.txt b/CMakeLists.txt
-index ac24bf41b..c22fb44dd 100644
---- a/CMakeLists.txt
-+++ b/CMakeLists.txt
-@@ -33,7 +33,7 @@ set(MBEDTLS_DIR ${CMAKE_CURRENT_SOURCE_DIR})
- option(USE_PKCS11_HELPER_LIBRARY "Build mbed TLS with the pkcs11-helper library." OFF)
- option(ENABLE_ZLIB_SUPPORT "Build mbed TLS with zlib library." OFF)
- 
--option(ENABLE_PROGRAMS "Build mbed TLS programs." ON)
-+option(ENABLE_PROGRAMS "Build mbed TLS programs." OFF)
- 
- option(UNSAFE_BUILD "Allow unsafe builds. These builds ARE NOT SECURE." OFF)
- option(MBEDTLS_FATAL_WARNINGS "Compiler warnings treated as errors" ON)
-@@ -47,7 +47,7 @@ string(REGEX MATCH "MSVC" CMAKE_COMPILER_IS_MSVC "${CMAKE_C_COMPILER_ID}")
- if(CMAKE_COMPILER_IS_MSVC)
-     option(ENABLE_TESTING "Build mbed TLS tests." OFF)
- else()
--    option(ENABLE_TESTING "Build mbed TLS tests." ON)
-+    option(ENABLE_TESTING "Build mbed TLS tests." OFF)
- endif()
- 
- # Warning string - created as a list for compatibility with CMake 2.8
-@@ -267,6 +267,9 @@ add_subdirectory(library)
- # target. This library of objects is used by tests and programs CMake files
- # to define the test executables.
- #
-+
-+if(0)
-+MESSAGE ("ENABLE_TESTING = "${ENABLE_TESTING} " ENABLE_PROGRAMS = " ${ENABLE_PROGRAMS})
- if(ENABLE_TESTING OR ENABLE_PROGRAMS)
-     file(GLOB MBEDTLS_TEST_FILES ${CMAKE_CURRENT_SOURCE_DIR}/tests/src/*.c ${CMAKE_CURRENT_SOURCE_DIR}/tests/src/drivers/*.c)
-     add_library(mbedtls_test OBJECT ${MBEDTLS_TEST_FILES})
-@@ -328,3 +331,4 @@ if(ENABLE_TESTING)
-                     ${CMAKE_CURRENT_BINARY_DIR}/DartConfiguration.tcl COPYONLY)
-     endif()
- endif()
-+endif()
