+++ port/port_common.h
@@ -1,0 +1,17 @@
+#ifndef _PORT_COMMON_H_
+#define _PORT_COMMON_H_
+
+/**
+  * ----------------------------------------------------------------------------------------------------
+  * Includes
+  * ----------------------------------------------------------------------------------------------------
+  */
+/* Common */
+#include "pico/stdlib.h"
+#include "pico/binary_info.h"
+#include "pico/critical_section.h"
+#include "hardware/spi.h"
+#include "hardware/dma.h"
+#include "hardware/clocks.h"
+
+#endif
