+++ port/version.h
@@ -1,0 +1,9 @@
+/**
+  * ----------------------------------------------------------------------------------------------------
+  * Version
+  * ----------------------------------------------------------------------------------------------------
+  */
+
+#define MAJOR_VERSION    2
+#define MINOR_VERSION    0
+#define PATCH_VERSION    0
